Edit DEPEND info for gnome-extra/timer-applet
Package gnome-extra/timer-applet DEPENDs on Python. The previous
declaration explicitely stated dependency on version 2.4 or 2.5 of Python
because of the need of manually pulling dev-python/elementtree if using
virtual/python:2.4. Now Python 2.6 is stable (at least on x86/amd64) and
the old declaration prevented python-2.5 from being removed. The package
seems to work with python-2.6, so I'm extending the declaration.

I do not use a DEPEND like '>=virtual/python:2.5' because I can't check if
timer-applet works with Python 3.1 too.