Freeze www-client/epiphany and dependencies
I'm tired of frequently bumping the beta Epiphany ebuild and its
dependencies (expecially net-libs/webkit-gtk). Therefore, I'm putting in
the repo the version I use right now of the browser and of all its
dependencies. This version has some... issues, but it's quite stable for my
needs. I'll remove these ebuilds as soon as the final version of Gnome 2.28
is released.