Remove net-im/pidgin ebuild
Remove an outdated net-im/pidgin ebuild, referring to a version of Pidgin
with known security problems. The ebuild in the overlay, compared to the
one in Portage, had a different fix w.r.t. the "famous" intltool problem;
but, after some benchmarks, it seems that the official way of fixing it
does not take more than a few seconds. Therefore, stick with the official
ebuild starting from the newest version of Pidgin.