Remove a check from the net-im/skype ebuild
Remove a USE dependency from the latest Skype ebuild. My guess is that the
Skype code that needed that dependency was removed some time ago, as I have
a qt-gui[-accessibility] now and Skype works flawlessly. The bug to force
on that USE was submitted years ago and is mentioned below.

Reference bug: https://bugs.gentoo.org/177817