Add examples for the Qt5 framework
Examples are currently not shipped with Qt5. An attempt was made to
add them at a package-level at https://bugs.gentoo.org/568838, and
later at https://github.com/gentoo/gentoo/pull/4541, and ultimately
https://github.com/gentoo/qt/pull/153. None of those were accepted for
apparent lack of interest.

At https://bugs.gentoo.org/669268, instead, a monolithic package with
all the examples was added, that is indeed much better for
out-of-the-box experience in qt-creator. This commit adds to the
overlay the ebuild attached to the latter bug, with some slight tweaks.