Fix linking problem with qt-creator-4.8.2 and sys-devel/clang:10
The original error was in the form:

  /usr/lib/gcc/x86_64-pc-linux-gnu/9.2.0/../../../../x86_64-pc-linux-gnu/bin/ld: cannot find -lclangFormat
  /usr/lib/gcc/x86_64-pc-linux-gnu/9.2.0/../../../../x86_64-pc-linux-gnu/bin/ld: cannot find -lclangToolingInclusions
  /usr/lib/gcc/x86_64-pc-linux-gnu/9.2.0/../../../../x86_64-pc-linux-gnu/bin/ld: cannot find -lclangToolingCore
  /usr/lib/gcc/x86_64-pc-linux-gnu/9.2.0/../../../../x86_64-pc-linux-gnu/bin/ld: cannot find -lclangRewrite
  /usr/lib/gcc/x86_64-pc-linux-gnu/9.2.0/../../../../x86_64-pc-linux-gnu/bin/ld: cannot find -lclangLex
  /usr/lib/gcc/x86_64-pc-linux-gnu/9.2.0/../../../../x86_64-pc-linux-gnu/bin/ld: cannot find -lclangBasic

Reported upstream as https://bugs.gentoo.org/730620

The idea behind the patch is borrowed from
https://bugs.gentoo.org/711788, that fixed the problem for versions of
qt-creator greater than 4.12.0.

Since waiting for the stabilization of dev-qt/qt-creator-4.12.3
(https://bugs.gentoo.org/720240) will take a long time, a different
patch to the stable qt-creator-4.8.2 is required; however,
clang_installation.pri is slightly different in the 4.8 branch, so a
slightly different patching is required.